Product Inbound Model Bug fix and Supplier Refund and Return services

+++ src/main/java/com/watersolution/inventory/component/management/supplier/refund/model/db/SupplierRefundInventory.java
@@ -34,7 +34,7 @@ public class SupplierRefundInventory extends Auditable {
     private Integer qty;
 
     @Column(name = "unitprice")
-    private BigDecimal unitprice;
+    private Double unitprice;
 
     public SupplierRefundInventory() {
